+++ src/main/java/com/example/springbootlabmessages/WebControllers/UploadController.java
@@ -1,8 +1,6 @@
-package com.example.springbootlabmessages.web;
+package com.example.springbootlabmessages.WebControllers;
 
 import com.example.springbootlabmessages.User.UserService;
-import jakarta.persistence.EntityManager;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Controller;
@@ -12,7 +10,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.swing.text.html.parser.Entity;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
