--- src/main/java/com/example/springbootlabmessages/web/WebController.java
@@ -1,247 +1,0 @@
-package com.example.springbootlabmessages.web;
-import com.example.springbootlabmessages.Message.CreateMessageFormData;
-import com.example.springbootlabmessages.Message.Message;
-import com.example.springbootlabmessages.Message.MessageService;
-import com.example.springbootlabmessages.Translation.Language;
-import com.example.springbootlabmessages.Translation.LanguageDTO;
-import com.example.springbootlabmessages.Translation.TranslationService;
-import com.example.springbootlabmessages.User.User;
-import com.example.springbootlabmessages.User.UserFormData;
-import com.example.springbootlabmessages.User.UserSearchDataDTO;
-import com.example.springbootlabmessages.User.UserService;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import jakarta.servlet.http.HttpSession;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@Controller
-public class WebController {
-
-    private int messagesPerLoad = 1;
-
-    private Long currentMessageToTranslateId;
-    private String currentTranslation;
-
-    private final MessageService messageService;
-    private final UserService userService;
-
-    private final TranslationService translationService;
-
-
-    public WebController(MessageService messageService, UserService userService, TranslationService translationService) {
-        this.messageService = messageService;
-        this.userService = userService;
-        this.translationService = translationService;
-    }
-
-    // CREATE MESSAGE
-
-    @GetMapping("/createmessage")
-    String home(OAuth2AuthenticationToken authentication, Model model) {
-        OAuth2User principal = authentication.getPrincipal();
-        model.addAttribute("formdata", new CreateMessageFormData());
-        model.addAttribute("principal", principal);
-        return "createmessage";
-    }
-
-    @PostMapping("/createmessage")
-    public String createMessage(CreateMessageFormData message, OAuth2AuthenticationToken authentication) {
-        OAuth2User principal = authentication.getPrincipal();
-        var auth = SecurityContextHolder.getContext().getAuthentication();
-        if (auth.getAuthorities().stream().findFirst().get().getAuthority().equals("OAUTH2_USER")) {
-            message.setUser(userService.findById(principal.getAttribute("id")));
-            var messageToSave = message.toEntity();
-            messageService.save(messageToSave);
-        }
-        return "redirect:/";
-    }
-
-
-
-    // MESSAGE FEED
-    @GetMapping("/")
-    String getMessages(Model model, Language language, LanguageDTO selectedLang, HttpSession session) {
-        //SECURITY
-        var auth = SecurityContextHolder.getContext().getAuthentication();
-        model.addAttribute("principal", auth);
-        //TRANSLATION
-        model.addAttribute("lang", language);
-        List<Language> languagesList = List.of(Language.values());
-        model.addAttribute("languagesList", languagesList);
-        model.addAttribute("selectedLang", selectedLang);
-        model.addAttribute("translatedMessage", currentTranslation);
-        model.addAttribute("currentMessageToTranslateId", currentMessageToTranslateId);
-        //SEARCH
-        model.addAttribute("userSearchData", new UserSearchDataDTO());
-
-        // IF AUTHENTICATED
-        if (auth.getAuthorities().stream().findFirst().get().getAuthority().equals("OAUTH2_USER")) {
-            var listOfMessages = messageService.get10Messages(messagesPerLoad);
-            model.addAttribute("listOfMessages", listOfMessages);
-            if(session.getAttribute("messageList") != null){
-                model.addAttribute("listOfMessages", session.getAttribute("messageList"));
-                session.removeAttribute("messageList");
-            }
-            return "messages";
-        // IF NOT AUTHENTICATED
-        } else {
-            var listOfMessages = messageService.get10PublicMessages(messagesPerLoad);
-            model.addAttribute("listOfMessages", listOfMessages);
-            if(session.getAttribute("messageList") != null){
-                model.addAttribute("listOfMessages", messageService.filterPublicMessages((List<Message>) session.getAttribute("messageList")));
-                session.removeAttribute("messageList");
-            }
-            return "allMessages";
-        }
-    }
-
-    @GetMapping("/loadMoreMessages")
-    public String loadMoreMessages(Model model, Language language, LanguageDTO selectedLang) {
-        var auth = SecurityContextHolder.getContext().getAuthentication();
-        model.addAttribute("lang", language);
-        model.addAttribute("selectedLang", selectedLang);
-
-        if (auth != null && auth.isAuthenticated()) {
-            List<Language> languagesList = List.of(Language.values());
-            model.addAttribute("languagesList", languagesList);
-            messagesPerLoad += 1;
-            var listOfMessages = messageService.get10Messages(messagesPerLoad);
-            model.addAttribute("listOfMessages", listOfMessages);
-        } else {
-            messagesPerLoad += 1;
-            var listOfMessages = messageService.get10PublicMessages(messagesPerLoad);
-            model.addAttribute("listOfMessages", listOfMessages);
-        }
-        return "redirect:/";
-    }
-
-
-
-
-// MYPAGE
-    @GetMapping("/mypage")
-    String mypage(@AuthenticationPrincipal OAuth2User principal, Model model) {
-        var auth = SecurityContextHolder.getContext().getAuthentication();
-        if(auth != null || auth.isAuthenticated()){
-            var user = userService.findById(principal.getAttribute("id"));
-            model.addAttribute("user", user);
-            model.addAttribute("userdata", new UserFormData());
-            model.addAttribute("principal", principal);
-            return "mypage";
-        }
-        else {
-            return "redirect:/login";
-        }
-
-    }
-
-    @PostMapping("/mypage")
-    String updateUser(@AuthenticationPrincipal OAuth2User principal, @ModelAttribute User user) {
-        var nyUser = userService.findById(principal.getAttribute("id"));
-        userService.updateUser(nyUser, user);
-        return "redirect:/mypage";
-    }
-
-
-    @GetMapping("/mymessages")
-    String myMessages(OAuth2AuthenticationToken authentication, Model model) {
-        OAuth2User principal = authentication.getPrincipal();
-        model.addAttribute("principal", principal);
-        OAuth2User oAuth2User = authentication.getPrincipal();
-        Long userId = userService.findById(oAuth2User.getAttribute("id")).getId();
-        List<Message> messageList = messageService.getAllMessagesByUser(userId);
-        model.addAttribute("messageList", messageList);
-        return "mymessages";
-    }
-
-
-    @GetMapping("/editmessage/{id}")
-    public String editMessage(Model model, @PathVariable Long id, OAuth2AuthenticationToken authentication) {
-        OAuth2User principal = authentication.getPrincipal();
-        model.addAttribute("principal", principal);
-        model.addAttribute("message", messageService.findById(id));
-        return "editmessage";
-    }
-
-    @PostMapping("/editmessage/{id}")
-    public String updateMessage(@AuthenticationPrincipal OAuth2User principal, @PathVariable Long id, @ModelAttribute Message message) {
-        var user = userService.findById(principal.getAttribute("id"));
-        var oldMessage = messageService.findById(id);
-        messageService.updateMessage(oldMessage, message, user);
-        return "redirect:/mymessages";
-    }
-
-
-    // TRANSLATION
-    @PostMapping("/translate/{messageId}")
-    public String translateMessage(@PathVariable Long messageId, @ModelAttribute LanguageDTO selectedLang, @ModelAttribute Message message) throws JsonProcessingException {
-        String messageToTranslate = messageService.getMessageById(messageId).getText();
-        currentTranslation = translationService.translate(messageToTranslate, selectedLang.getLangCode());
-        currentMessageToTranslateId = messageId;
-        return "redirect:/";
-    }
-
-
-
-
-    // SEARCH
-@GetMapping("/result")
-    public String result(Model model, @ModelAttribute LanguageDTO selectedLang, Language language, HttpSession session) {
-        //SECURITY
-        var auth = SecurityContextHolder.getContext().getAuthentication();
-        //TRANSLATION
-        model.addAttribute("lang", language);
-        List<Language> languagesList = List.of(Language.values());
-        model.addAttribute("languagesList", languagesList);
-        model.addAttribute("selectedLang", selectedLang);
-        model.addAttribute("translatedMessage", currentTranslation);
-        model.addAttribute("currentMessageToTranslateId", currentMessageToTranslateId);
-
-        if (auth.getAuthorities().stream().findFirst().get().getAuthority().equals("OAUTH2_USER")) {
-            var listOfMessages = messageService.get10Messages(messagesPerLoad);
-            model.addAttribute("listOfMessages", listOfMessages);
-            if(session.getAttribute("messageList") != null){
-                model.addAttribute("listOfMessages", session.getAttribute("messageList"));
-            }
-        }
-        else {
-            var listOfMessages = messageService.get10PublicMessages(messagesPerLoad);
-            model.addAttribute("listOfMessages", listOfMessages);
-            if(session.getAttribute("messageList") != null){
-                model.addAttribute("listOfMessages", messageService.filterPublicMessages((List<Message>) session.getAttribute("messageList")));
-            }
-        }
-        return "results";
-        }
-
-    @PostMapping("/translateSearch/{messageId}")
-    public String translateSearch(@PathVariable Long messageId, @ModelAttribute LanguageDTO selectedLang, @ModelAttribute Message message) throws JsonProcessingException {
-        String messageToTranslate = messageService.getMessageById(messageId).getText();
-        currentTranslation = translationService.translate(messageToTranslate, selectedLang.getLangCode());
-        currentMessageToTranslateId = messageId;
-        return "redirect:/result";
-    }
-
-
-
-    @PostMapping("/search")
-    public String searchMessages(@ModelAttribute UserSearchDataDTO searchText, HttpSession session) {
-        List<Message> listOfMessagesBasedOnSearch = messageService.findAllByUserName(searchText.getSearchText());
-        session.setAttribute("messageList", listOfMessagesBasedOnSearch);
-        return "redirect:/result";
-    }
-
-
-}
-
-
-
-
-
