+++ src/main/resources/templates/messages.html
@@ -32,8 +32,10 @@
     <div class="card-body">
         <h5 class="center-message" th:text="${message?.title}"></h5>
         <p class="center-message" th:text="${message?.text}"></p>
+        <div style="margin-top: 10px">
         <p class="center-message" th:if="${translatedMessage != null && currentMessageToTranslateId == message.id}" th:text="'Translation: ' + ${translatedMessage}"></p>
-        <div class="user-info">
+        </div>
+            <div class="user-info">
             <p class="message" th:text="|#{Author}: ${message?.user.username}|"></p>
             <p class="message" th:text="|#{PublishedTime}: ${message?.createdAt}|"></p>
         </div>
