+++ src/main/resources/templates/mymessages.html
@@ -58,13 +58,18 @@
     }
 
     .message {
-      font-size: 18px;
+      font-size: 14px;
       text-align: center;
+      display: flex;
+      flex-direction: row;
+      justify-content: space-between;
+      margin-bottom: 14px;
     }
 
     .user-info {
       display: flex;
       justify-content: space-evenly;
+      width: 100%;
     }
 
     #english-btn {
@@ -144,8 +149,20 @@
       font-size: 20px;
       font-weight: bold;
       cursor: pointer;
-      position: absolute;
+      margin-bottom: 10px;
     }
+    .card-body{
+      display: flex;
+      flex-direction: column;
+      justify-content: space-evenly;
+      align-items: center;
+      border-top: solid black 1px;
+      border-bottom: solid black 1px;
+      width : 70%;
+      margin-left: auto;
+      margin-right: auto;
+    }
+
 
   </style>
 </head>
@@ -164,16 +181,18 @@
     <h1 class="site-title">My messages</h1>
   </div>
 
-  <div th:each="message : ${messageList}" class="card mb-3">
-    <div class="card-body">
-      <h5 class="center title" th:text="${ message?.title }"></h5>
-      <p id="center message" th:text="${ message?.text}"></p>
+  <div th:each="message : ${messageList}" class="card-body">
+
+      <h5 class="center title" th:text="'Title: ' +  ${ message?.title }" style="text-decoration: underline; font-size: 16px" ></h5>
+      <p id="message" th:text= "${ message?.text}"></p>
       <div class="user-info">
         <p class="message" th:text="#{Author} + ': ' + ${message?.user.username}"></p>
         <p class="message" th:text="#{PublishedTime}+ ': ' + ${message?.createdAt}"></p>
-      </div>
-      <a th:href="@{'/editmessage/' + ${message.id}}" class="edit-btn" id="edit-btn">Edit</a>
+      </div >
+        <a th:href="@{'/editmessage/' + ${message.id}}" class="edit-btn" id="edit-btn">Edit</a>
+
+
+
     </div>
-  </div>
 </body>
 </html>
