+++ src/main/resources/templates/results.html
@@ -11,7 +11,7 @@
     <a href="http://localhost:8080/?lang=sv" id="swedish-btn">Swedish</a>
 </div>
 <div class="header-buttons">
-    <a class="header-button" href="http://localhost:8080/login" >Log in</a>
+
 </div>
 
 <div>
@@ -28,8 +28,11 @@
     <div class="card-body">
         <h5 class="center-message" th:text="${message?.title}"></h5>
         <p class="center-message" th:text="${message?.text}"></p>
-        <p class="center-message" th:if="${translatedMessage != null && currentMessageToTranslateId == message.id}" th:text="'Translation: ' + ${translatedMessage}"></p>
-        <div class="user-info">
+<div style="margin-top: 10px">
+    <p class="center-message" th:if="${translatedMessage != null && currentMessageToTranslateId == message.id}" th:text="'Translation: ' + ${translatedMessage}"></p>
+
+</div>
+              <div class="user-info">
             <p class="message" th:text="|#{Author}: ${message?.user.username}|"></p>
             <p class="message" th:text="|#{PublishedTime}: ${message?.createdAt}|"></p>
         </div>
